Phase 1: fix correctness and safety bugs
- price_notify: set priceNotifyInit on CONNECTED/DISCONNECTED so the
  watchdog branch in monitorDataConnections actually runs
- block_notify: clear notifyInit in stop() and on DISCONNECTED so
  isInitialized() reflects reality after restart
- config: return early from getFsRev() when LittleFS.open fails instead
  of reading from an invalid File handle
- led_handler: fix frontlightSetBrightness off-by-one that wrote one
  PCA9685 channel past the array
- nostrdisplay_handler: widen parseZapNotify amount from uint16_t to
  uint64_t so large zaps no longer truncate
- button_handler: stop holding mcpMutex across delay(1); snapshot MCP
  state under the lock, release before dispatching and INT polling
- timers: cache bitaxe + mining pool task handles in module-level
  volatile statics so minuteTimerISR no longer reaches into flash-resident
  singleton accessors from IRAM; null-check taskScreenRotateTaskHandle
  in screenRotateTimerISR
- webserver: remove Sec-Fetch-Mode onNotFound shortcut (was masking
  404s); require HTTP auth on status / system_status / identify /
  lights / frontlight status / DND status; enable auth on the SSE
  EventSource when http auth is enabled
- button_handler.hpp: drop dead double-click / long-press API and the
  ButtonState fields that only fed it

+++ lib/btclock/nostrdisplay_handler.cpp
@@ -1,6 +1,6 @@
 #include "nostrdisplay_handler.hpp"
 
-std::array<std::string, NUM_SCREENS> parseZapNotify(std::uint16_t amount, bool withSatsSymbol)
+std::array<std::string, NUM_SCREENS> parseZapNotify(std::uint64_t amount, bool withSatsSymbol)
 {
     // Initialize defensively so it works for any NUM_SCREENS value (previously
     // a fixed 7-element initializer assumed NUM_SCREENS == 7).
